temporary move snapDB to consensus layer, clique and other consensus could use it too;
